pyproject.toml
[project]
name = "skyloom"
version = "1.0.0"
description = "A light weight weather dashboard application"
requires-python = ">=3.12"
dependencies = [
    "pyside6>=6.10.1",
    "pyside6-fluent-widgets>=1.10.5",
    "build>=1.4.0",
    "pip-autoremove",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1"
]
